[How] 1. Add a Match case check box    1.1 Create UI    1.2 Create IF of CLogFileLoader    1.3 TODO: need to find wcsistr function 2. refactor the usage of PROPINFO    2.1 constructor    2.2 just assigned

+++ NGLogViewer/main.cpp
@@ -66,7 +66,7 @@ int TestCRegSetting()
 {
 	class CRegSetting cRegSetting(L"SOFTWARE\\NG1\\BBB");
 	DWORD a=0;
-	cRegSetting.WriteRegKey(L"SettingA", 5);
+	cRegSetting.WriteRegKey(L"SettingA", (DWORD)5);
 	cRegSetting.ReadRegKey(L"SettingA", a);
 	wchar_t wstr[128] = {0};
 	cRegSetting.WriteRegKey(L"SettingString", L"abcd");
